Misc PR adjustments + cleanup; language translation integration; WIP.
