+++ messages/ewe-BJ.json
@@ -194,6 +194,7 @@
     "SEARCH_LOADING": "Gɔgɔme nyateƒe kplɔ...",
     "REPLY": "Ɖe si",
     "SAVE": "Ŋutɔ",
+    "CONFIRM": "Tsi me",
     "CONFIRM_DIALOG": "Ɖe si kple alo ɖe be si ɖe ɖe ɖe siaɖiwo. Wò be wo kple agbe?",
     "USER_INFORMATION": {
       "PI_USERNAME_LABEL": "Nuxlɔ̃tɔ ŋkɔ Pi",
