+++ messages/fon-BJ.json
@@ -194,6 +194,7 @@
     "SEARCH_LOADING": "Biba ɖo yiyi wɛ...",
     "REPLY": "Xósin",
     "SAVE": "Hɛn",
+    "CONFIRM": "Ganmɛ",
     "CONFIRM_DIALOG": "A ɖo ɖyɔɖyɔ e ma ko hɛn ɖ'ayi ǎ lɛ. A jlo na gosin nugbo a?",
     "USER_INFORMATION": {
       "PI_USERNAME_LABEL": "Pi nyikɔ mitɔn",
