+++ messages/hau-NG.json
@@ -73,6 +73,7 @@
       "SELLER_DETAILS_LABEL": "Bayanin Mai Sayarwa",
       "SELLER_ADDRESS_POSITION_LABEL": "Adireshin Ko Matsayin Mai Sayarwa",
       "LEAVE_A_REVIEW_MESSAGE": "Bar Saƙon Bita",
+      "FACE_SELECTION_REVIEW_MESSAGE": "Zaɓi fuskar da ke nuna yadda kake ji game da wannan mai siyarwa na sama",	
       "ADDITIONAL_COMMENTS_PLACEHOLDER": "Shigar da karin sharhi a nan",
       "FEEDBACK_PHOTO_UPLOAD_LABEL": "Loda Hoto Martani",
       "REVIEWS_SUMMARY_LABEL": "Takaitaccen Bayanin Sake Dubawa",
@@ -102,7 +103,7 @@
       "REPLY_TO_REVIEW_STATIC_HEADER": "Amsa Bita",
       "REPLY_TO_REVIEW_SUBHEADER": "Bitar da kake maida martani",
       "GIVE_REPLY_TO_REVIEW_SUBHEADER": "Bayar da amsa ga bita",
-      "REPLY_TO_REVIEW_HEADER": "Amsa wa bita",
+      "REPLY_TO_REVIEW_HEADER": "Amsa wa bita da aka bar wa {seller_id}",
       "BY_REVIEWER": "Ta {buyer_id}",
       "REPLY_TO_REVIEW_MESSAGE": "Bar amsar ku ga bita ta sama",
       "FACE_SELECTION_REVIEW_MESSAGE": "Zaɓi fuskar da ke nuna yadda kake ji game da mu'amalarka da wannan jagora",
@@ -193,6 +194,7 @@
     "SEARCH_LOADING": "Ana ɗora bincike...",
     "REPLY": "Amsa",
     "SAVE": "Ajiye",
+    "CONFIRM": "Tabbatar",
     "CONFIRM_DIALOG": "Kuna da canje-canje marasa ajiya. Da gaske kuna son barin?",
     "USER_INFORMATION": {
       "PI_USERNAME_LABEL": "Sunan mai amfani na Pi",
