+++ src/components/shared/confirm.tsx
@@ -1,13 +1,16 @@
+import { useTranslations } from "next-intl";
+import { useRouter } from 'next/navigation';
 import React, { SetStateAction } from 'react';
 import { createPortal } from 'react-dom';
 import { IoMdClose } from 'react-icons/io';
-import { useRouter } from 'next/navigation';
+
+const t = useTranslations();
 
 const ConfirmDialog = ({ show, onClose, message, url }:any) => {
   const router = useRouter();
     
   const handleClicked = () => {
-    router.push(url)
+    router.push(url);
   }
 
   if (!show) return null;
@@ -28,7 +31,7 @@ const ConfirmDialog = ({ show, onClose, message, url }:any) => {
               onClick={handleClicked}
               className="px-4 py-2 bg-primary text-secondary text-xl rounded-md"
             >
-              Confirm
+              {t('SHARED.CONFIRM')}
             </button>
           </div>
         </div>
@@ -37,8 +40,6 @@ const ConfirmDialog = ({ show, onClose, message, url }:any) => {
     document.body
   );
 };
-
-export default ConfirmDialog;
 
 export const ConfirmDialogX = ({ toggle, handleClicked, message }:any) => {
   return createPortal(
@@ -57,7 +58,7 @@ export const ConfirmDialogX = ({ toggle, handleClicked, message }:any) => {
               onClick={handleClicked}
               className="px-4 py-2 bg-primary text-secondary text-xl rounded-md"
             >
-              Confirm
+              {t('SHARED.CONFIRM')}
             </button>
           </div>
         </div>
@@ -73,23 +74,25 @@ export const Notification:React.FC<{
   setShowDialog: React.Dispatch<SetStateAction<boolean>>;
 }> = ({ message, showDialog, setShowDialog }) => {
 
-  const onClose = ()=>{
-    setShowDialog(false)
+  const onClose = () => {
+    setShowDialog(false);
   }
 
-    return (
-      <div className={`absolute inset-0 flex items-center justify-center bg-black bg-opacity-50 z-20 ${!showDialog? 'hidden': ''}`}>
-        <div className="bg-white rounded-lg p-6 w-full max-w-md mx-6 sm:mx-auto relative">
-          <button
+  return (
+    <div className={`absolute inset-0 flex items-center justify-center bg-black bg-opacity-50 z-20 ${!showDialog? 'hidden': ''}`}>
+      <div className="bg-white rounded-lg p-6 w-full max-w-md mx-6 sm:mx-auto relative">
+        <button
           onClick={onClose}
           className="absolute top-3 right-3"
-          >
+        >
           <IoMdClose size={24} className="text-gray-600 hover:text-gray-800" />
-          </button>
-          <div className="text-center mt-5">
+        </button>
+        <div className="text-center mt-5">
           <p className="text-lg mb-4">{message}</p>
-          </div>
         </div>
       </div>
-    )
-  }
+    </div>
+  )
+}
+
+export default ConfirmDialog;
