+++ src/services/sellerApi.ts
@@ -99,33 +99,10 @@ export const registerSeller = async (formData: FormData) => {
   }
 };
 
-// Add or update seller item
-export const addOrUpdateSellerItem = async (formData: FormData) => {
-  try {
-    logger.info('Creating or updating seller registration with formData..');
-    const headers = getMultipartFormDataHeaders();
-    console.log('Form data being sent:', Object.fromEntries(formData.entries()));
-    const response = await axiosClient.put('/sellers/item/add', formData, { headers });
-
-    if (response.status === 200) {
-      logger.info(`Add or update seller item successful with Status ${response.status}`, {
-        data: response.data
-      });
-      return response.data;
-    } else {
-      logger.error(`Add or update seller item failed with Status ${response.status}`);
-      return null;
-    }
-  } catch (error) {
-    logger.error('Add or update seller item encountered an error:', error);
-    throw new Error('Failed to modify seller. Please try again later.');
-  }
-};
-
-//fetch all items for a seller
+// Fetch all seller items associated with the seller
 export const fetchSellerItems = async (sellerId: string) => {
   try {
-    logger.info(`Fetching seller items with ID: ${sellerId}`);
+    logger.info(`Fetching seller items associated with sellerID: ${sellerId}`);
     const response = await axiosClient.get(`/sellers/item/${sellerId}`);
     if (response.status === 200) {
       logger.info(`Fetch seller items successful with Status ${response.status}`, {
@@ -142,10 +119,33 @@ export const fetchSellerItems = async (sellerId: string) => {
   }
 };
 
-//delee an item for a seller
+// Add or update seller item
+export const addOrUpdateSellerItem = async (formData: FormData) => {
+  try {
+    logger.info('Creating or updating seller item with formData..');
+    const headers = getMultipartFormDataHeaders();
+
+    const response = await axiosClient.put('/sellers/item/add', formData, { headers });
+
+    if (response.status === 200) {
+      logger.info(`Add or update seller item successful with Status ${response.status}`, {
+        data: response.data
+      });
+      return response.data;
+    } else {
+      logger.error(`Add or update seller item failed with Status ${response.status}`);
+      return null;
+    }
+  } catch (error) {
+    logger.error('Add or update seller item encountered an error:', error);
+    throw new Error('Failed to add or update seller item. Please try again later.');
+  }
+};
+
+// Delete a seller item
 export const deleteSellerItem = async (itemId: string) => {
   try {
-    logger.info(`deleting seller items with ID: ${itemId}`);
+    logger.info(`Deleting seller item with itemID: ${itemId}`);
     const response = await axiosClient.delete(`/sellers/item/delete/${itemId}`);
     if (response.status === 200) {
       logger.info(`Delete seller item successful with Status ${response.status}`, {
@@ -153,12 +153,12 @@ export const deleteSellerItem = async (itemId: string) => {
       });
       return response.data;
     } else {
-      logger.error(`Delete seller items failed with Status ${response.status}`);
+      logger.error(`Delete seller item failed with Status ${response.status}`);
       return null;
     }
   } catch (error) {
-    logger.error('Delete seller items encountered an error:', error);
-    throw new Error('Failed to Delete seller items. Please try again later.');
+    logger.error('Delete seller item encountered an error:', error);
+    throw new Error('Failed to delete seller item. Please try again later.');
   }
 };
